app-drones/requirements.txt
amqp==5.2.0
billiard==4.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Flask==3.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.5
pika==1.2.0
prompt-toolkit==3.0.43
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
azure-storage-blob==12.9.0
psycopg2-binary == 2.9.9